marketplace-analytics/requirements.txt
kafka-python==2.0.2
faust-streaming==0.11.11
psycopg2-binary==2.9.9
